develop market section and romove some duplicate queries on sub-navigation bar

+++ database/migrations/2022_07_28_075842_create_teams_table.php
@@ -13,14 +13,16 @@ return new class extends Migration
      */
     public function up()
     {
-        Schema::create('teams', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('user_id')->constrained()->onDelete('cascade');
-            $table->foreignId('league_id')->constrained()->onDelete('cascade');
-            $table->string('name', 25);
-            $table->string('stadium', 25);
-            $table->timestamps();
-        });
+        if(!Schema::hasTable('teams')){
+            Schema::create('teams', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('user_id')->constrained()->onDelete('cascade');
+                $table->foreignId('league_id')->constrained('user_settings')->onDelete('cascade');
+                $table->string('name', 25);
+                $table->string('stadium', 25);
+                $table->timestamps();
+            });
+        }
     }
 
     /**
